+++ CHANGELOG.markdown
@@ -1,5 +1,10 @@
 # OpenVBX Change Log
 
+## Next
+
+- Send proper JSON header during install steps.
+
+
 ## OpenVBX 1.2.12
 
 - Fix cache table `value` field to be `mediumblob` instead of `text`. Fixes an issue where large lists of phone numbers (300+) would overflow the value field and corrupt the cached data.